chore: refactor meta_client, decouple interface definitions from implementations
